% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartocolors.R
\docType{data}
\name{cartocolors}
\alias{cartocolors}
\title{CARTOColors database}
\format{A data frame with 32 rows and 12 variables:
\itemize{
\item Name: palette name
\item Type: palette type
\item n2 to n11: list of the color values depending on the number of colors used (between two and twelve)
}}
\usage{
cartocolors
}
\description{
A dataset containing the information about the CARTOcolor palettes.
Source: https://github.com/CartoDB/CartoColor/blob/master/cartocolor.js.
}
\keyword{datasets}
