% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_carto_all.R
\name{display_carto_all}
\alias{display_carto_all}
\title{CARTOcolor palettes viewer}
\usage{
display_carto_all(n = NULL, type = "all")
}
\arguments{
\item{n}{Number of different colors in the palette, minimum depending on palette (2 or 3), maximum depending on palette (7 or 11)}

\item{type}{Type of the palette, could be "quantitative", "diverging", "qualitative", "aggregation", or "all"}
}
\value{
A character vector
}
\description{
Displays a set of color palettes from CARTOcolor
}
\examples{
display_carto_all()
display_carto_all(type = "quantitative")
display_carto_all(type = c("diverging", "qualitative", "aggregation"))
display_carto_all(3, type = "quantitative")
display_carto_all(7, type = "quantitative")

}
