% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_carto_pal.R
\name{display_carto_pal}
\alias{display_carto_pal}
\title{CARTOcolor palette viewer}
\usage{
display_carto_pal(n, name)
}
\arguments{
\item{n}{Number of different colors in the palette, minimum depending on palette (2 or 3), maximum depending on palette (7 or 11)}

\item{name}{A palette name}
}
\value{
A character vector
}
\description{
Displays a color palette from CARTOcolor
}
\examples{
display_carto_pal(7, "Burg")

}
