% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacartocolors.R
\docType{data}
\name{metacartocolors}
\alias{metacartocolors}
\title{CARTOColors database metadata}
\format{A data frame with 32 rows and 12 variables:
\itemize{
\item Name: palette name
\item Type: palette type
\item Min_n: minimal number of colors in a given palette
\item Max_n: maximal number of colors in a given palette
}}
\usage{
metacartocolors
}
\description{
A dataset containing the information about the CARTOcolor palettes.
Source: https://github.com/CartoDB/CartoColor/blob/master/cartocolor.js.
}
\keyword{datasets}
