% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto_pal.R
\name{carto_pal}
\alias{carto_pal}
\title{CARTOcolor palettes}
\usage{
carto_pal(n = NULL, name)
}
\arguments{
\item{n}{Number of different colors in the palette, minimum depending on palette (2 or 3), maximum depending on palette (7 or 11)}

\item{name}{A palette name}
}
\value{
A character vector
}
\description{
Creates a color palette from CARTOcolor
}
\examples{
my_colors = carto_pal(7, "Burg")

}
